add:api product

+++ app/Http/Controllers/UserController.php
@@ -2,21 +2,22 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Product;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
-class ProductController extends Controller
+use function PHPUnit\Framework\returnSelf;
+
+class UserController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $data = Product::latest()->get();
-
+        $data = User::latest()->get();
         return response()->json([
-            'msg'=>'success',
+            'msg'=>'List Users',
             'data'=>$data
         ],200);
     }
@@ -35,23 +36,21 @@ class ProductController extends Controller
     public function store(Request $request)
     {
         $validate = $request->validate([
-            'name'=>'required|unique:products',
-            'description'=>'required',
-            'stok'=>'required|numeric',
-            'price'=>'required|numeric',
+            'name'=>'required',
+            'email'=>'required|email|unique:users',
+            'password'=>'required|min:6',
             'image'=>'mimes:png,jpg|required|max:2048'
         ]);
 
-        if($request->hasFile('image'))
-        {
-            $image = $request->file('image');
-            $image->storeAs('public/products',$image->hashName());
-        }
-        $validate['image'] = $image->hashName();
-        $data = Product::create($validate);
+        $image = $request->file('image');
+        $image->storeAs('public/users',$image->hashName());
+
+        $validate['image']=$image->hashName();
+
+        $data = User::create($validate);
 
         return response()->json([
-            'msg'=>'success created data products',
+            'msg'=>'success created user',
             'data'=>$data
         ],200);
     }
@@ -61,27 +60,26 @@ class ProductController extends Controller
      */
     public function show($id)
     {
-        $data = Product::find($id);
+        $data = User::find($id);
 
         if(!$data){
             return response()->json([
-                'msg'=>'Data Not Found',
+                'msg'=>'User Not Found'
             ],404);
-        }
+        };
 
         return response()->json([
-            'msg'=>'Detail Product',
+            'msg'=>'User Detail',
             'data'=>$data
         ],200);
-
     }
 
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Product $product)
+    public function edit(User $user)
     {
-        //
+
     }
 
     /**
@@ -89,37 +87,34 @@ class ProductController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $data = Product::find($id);
+        $data = User::find($id);
 
-        if(!$data)
-        {
+        if(!$data){
             return response()->json([
-                'msg'=>'Data Not Found',
+                'msg'=>'User Not Found'
             ],404);
-        }
+        };
 
         $validate = $request->validate([
-            'name'=>'required|unique:products,name,id,'.$id,
-            'description'=>'required',
-            'stok'=>'required|numeric',
-            'price'=>'required|numeric',
+            'name'=>'required',
+            'email'=>'required|email|unique:users,email,id,'.$id,
+            'password'=>'required|min:6',
             'image'=>'mimes:png,jpg|nullable|max:2048'
         ]);
 
         if($request->hasFile('image')){
-            Storage::delete('public/products/'.basename($data->image));
-
+            Storage::delete('public/users/'.basename($data->image));
 
             $image = $request->file('image');
-            $image->storeAs('public/products',$image->hashName());
+            $image->storeAs('public/users',$image->hashName());
 
-            $validate['image']= $image->hashName();
-        }
+            $validate['image']=$image->hashName();
+        };
 
-        $datas=$data->update($validate);
+        $datas = $data->update($validate);
 
         return response()->json([
-            'msg'=>'success update data product',
+            'msg'=>'success updated data',
             'data'=>$datas
         ],200);
     }
@@ -129,12 +124,21 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
-        $data=Product::find($id);
+        $data = User::find($id);
 
-        Storage::delete('public/products/'.basename($data->image));
+        if(!$data){
+            return response()->json([
+                'msg'=>'User Not Found'
+            ],404);
+        };
+
+        Storage::delete('public/users/'.basename($data->image));
+
         $data->delete();
+
         return response()->json([
-            'msg'=>"Success Deleted Product"
+            'msg'=>'success deleted user'
         ],200);
+
     }
 }
